--- NIPTool/server/load/load/api/api_v1/api.py
@@ -1,8 +1,0 @@
-from fastapi import FastAPI
-
-from NIPTool.server.load.api.api_v1.endpoints import load
-
-
-app = FastAPI()
-
-app.include_router(load.router, prefix="/api/v1/load", tags=["load"])
