--- NIPTool/server/load/load/api/api_v1/endpoints/__init__.py
@@ -1,1 +1,0 @@
-
